refactor: split sample sources config
